refactor: SignupInputType -> SignupValidator로 네이밍 수정
